requirements.txt
google-genai
SpeechRecognition
pyttsx3
pyaudio
python-dotenv